lab3/requirements.txt
numpy==1.25.2
torch==2.1.2
transformers==4.41.1
sentence-transformers==2.7.0
faiss-cpu==1.7.4
streamlit
rapidfuzz
pandas
pdfplumber
yake
bs4
feedparser
stanza
nltk
matplotlib
seaborn
scipy
textblob